requirements.txt
streamlit>=1.24
openai>=1.0.0
langchain>=0.0.300
playwright>=1.35.0
typing-extensions>=4.5.0
aiohttp>=3.8.4
python-dotenv>=1.0.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
googlemaps>=4.10.0
google-search-results>=2.4.0
pydantic>=2.0.0
